+++ tests/test_autorun.py
@@ -10,7 +10,6 @@ from immutable import Immutable
 
 from redux.basic_types import (
     BaseAction,
-    BaseEvent,
     CompleteReducerResult,
     CreateStoreOptions,
     FinishAction,
@@ -55,8 +54,11 @@ def reducer(
     return state
 
 
+StoreType = Store[StateType, Action, FinishEvent]
+
+
 @pytest.fixture()
-def store() -> Generator[Store[StateType, Action, FinishEvent], None, None]:
+def store() -> Generator[StoreType, None, None]:
     store = Store(reducer, options=CreateStoreOptions(auto_init=True))
     yield store
     store.dispatch(IncrementAction())
@@ -65,10 +67,7 @@ def store() -> Generator[Store[StateType, Action, FinishEvent], None, None]:
     store.dispatch(FinishAction())
 
 
-def test_general(
-    store_snapshot: StoreSnapshotContext,
-    store: Store[StateType, Action, BaseEvent],
-) -> None:
+def test_general(store_snapshot: StoreSnapshotContext, store: StoreType) -> None:
     store_snapshot.set_store(store)
 
     @store.autorun(lambda state: state.value)
@@ -79,7 +78,7 @@ def test_general(
 
 def test_ignore_attribute_error_in_selector(
     store_snapshot: StoreSnapshotContext,
-    store: Store[StateType, Action, BaseEvent],
+    store: StoreType,
 ) -> None:
     store_snapshot.set_store(store)
 
@@ -90,7 +89,7 @@ def test_ignore_attribute_error_in_selector(
 
 def test_ignore_attribute_error_in_comparator(
     store_snapshot: StoreSnapshotContext,
-    store: Store[StateType, Action, BaseEvent],
+    store: StoreType,
 ) -> None:
     store_snapshot.set_store(store)
 
@@ -102,10 +101,7 @@ def test_ignore_attribute_error_in_comparator(
         pytest.fail('This should never be called')
 
 
-def test_with_old_value(
-    store_snapshot: StoreSnapshotContext,
-    store: Store[StateType, Action, BaseEvent],
-) -> None:
+def test_with_old_value(store_snapshot: StoreSnapshotContext, store: StoreType) -> None:
     store_snapshot.set_store(store)
 
     @store.autorun(lambda state: state.value)
@@ -116,7 +112,7 @@ def test_with_old_value(
 
 def test_with_comparator(
     store_snapshot: StoreSnapshotContext,
-    store: Store[StateType, Action, BaseEvent],
+    store: StoreType,
 ) -> None:
     store_snapshot.set_store(store)
 
@@ -131,7 +127,7 @@ def test_with_comparator(
 
 def test_with_comparator_and_old_value(
     store_snapshot: StoreSnapshotContext,
-    store: Store[StateType, Action, BaseEvent],
+    store: StoreType,
 ) -> None:
     store_snapshot.set_store(store)
 
@@ -144,10 +140,7 @@ def test_with_comparator_and_old_value(
         return value - (old_value or 0)
 
 
-def test_value_property(
-    store_snapshot: StoreSnapshotContext,
-    store: Store[StateType, Action, BaseEvent],
-) -> None:
+def test_value_property(store_snapshot: StoreSnapshotContext, store: StoreType) -> None:
     store_snapshot.set_store(store)
 
     @store.autorun(lambda state: state.value)
@@ -164,10 +157,7 @@ def test_value_property(
     render.subscribe(check)
 
 
-def test_callability(
-    store_snapshot: StoreSnapshotContext,
-    store: Store[StateType, Action, BaseEvent],
-) -> None:
+def test_callability(store_snapshot: StoreSnapshotContext, store: StoreType) -> None:
     store_snapshot.set_store(store)
 
     @store.autorun(lambda state: state.value)
@@ -181,10 +171,7 @@ def test_callability(
     store.subscribe(check)
 
 
-def test_subscription(
-    store_snapshot: StoreSnapshotContext,
-    store: Store[StateType, Action, BaseEvent],
-) -> None:
+def test_subscription(store_snapshot: StoreSnapshotContext, store: StoreType) -> None:
     store_snapshot.set_store(store)
 
     @store.autorun(lambda state: state.value)
@@ -197,10 +184,7 @@ def test_subscription(
     render.subscribe(reaction, initial_run=True)
 
 
-def test_unsubscription(
-    store_snapshot: StoreSnapshotContext,
-    store: Store[StateType, Action, BaseEvent],
-) -> None:
+def test_unsubscription(store_snapshot: StoreSnapshotContext, store: StoreType) -> None:
     store_snapshot.set_store(store)
 
     @store.autorun(lambda state: state.value)
@@ -214,10 +198,7 @@ def test_unsubscription(
     unsubscribe()
 
 
-def test_repr(
-    store_snapshot: StoreSnapshotContext,
-    store: Store[StateType, Action, BaseEvent],
-) -> None:
+def test_repr(store_snapshot: StoreSnapshotContext, store: StoreType) -> None:
     store_snapshot.set_store(store)
 
     @store.autorun(lambda state: state.value)